treewide: Improve nonfree CI and targets, remove WIP tests (#11)
* target/sim: Parametrize tb top and bootmode selection

* Control simulation key parameters to ease CI and user simulation
isntead of hardcoded values for tb top and bootmode

* sw/tests: Add preliminary bootmode check test

* The test reads the bootmode under use and prints it. Bootmode correct
working is tested by the code itself, i.e. if the selected bootmode
works, the code will be preloaded correctly in SRAM and executed.

* sw/tests: Set correct frequency for RTL sim

* TODO: Improve the setup. There should be no manual intervention from
the user in the application code when compiling for simulation or fpga.
Move the configuration one layer above

* sw/tests: Remove non functional test

* target/xilinx: Add timing closure check

* Fail if timing is not closed

* Makefile: Update pointer to nonfree repository
